+++ exec/command_test.go
@@ -1,4 +1,4 @@
-package types
+package exec
 
 import (
 	"bytes"
@@ -20,7 +20,7 @@ func TestCmdRun(t *testing.T) {
 		{
 			name: "trace enabled",
 			cmd: &Cmd{
-				Trace: boolPtr(true),
+				Trace: true,
 				Cmd: &exec.Cmd{
 					Path: "/usr/bin/echo",
 					Args: []string{"echo", "hello"},
@@ -30,19 +30,20 @@ func TestCmdRun(t *testing.T) {
 			wantStdout: "hello\n",
 		},
 		{
-			name: "private output",
+			name: "trace disabled",
 			cmd: &Cmd{
-				Private: true,
+				Trace: false,
 				Cmd: &exec.Cmd{
 					Path: "/usr/bin/echo",
 					Args: []string{"echo", "hello"},
 				},
 			},
-			wantTrace: "+ echo hello\n",
+			wantStdout: "hello\n",
 		},
 		{
 			name: "custom env",
 			cmd: &Cmd{
+				Trace: true,
 				Cmd: &exec.Cmd{
 					Path: "/bin/sh",
 					Args: []string{"sh", "-c", "echo $TEST"},
@@ -53,20 +54,9 @@ func TestCmdRun(t *testing.T) {
 			wantStdout: "1\n",
 		},
 		{
-			name: "custom stdout",
-			cmd: &Cmd{
-				Cmd: &exec.Cmd{
-					Path:   "/bin/sh",
-					Args:   []string{"sh", "-c", "echo hello"},
-					Stdout: new(bytes.Buffer),
-				},
-			},
-			wantTrace:  "+ sh -c echo hello\n",
-			wantStdout: "hello\n",
-		},
-		{
 			name: "custom stderr",
 			cmd: &Cmd{
+				Trace: true,
 				Cmd: &exec.Cmd{
 					Path:   "/bin/sh",
 					Args:   []string{"sh", "-c", "echo error >&2"},
@@ -75,6 +65,16 @@ func TestCmdRun(t *testing.T) {
 			},
 			wantTrace:  "+ sh -c echo error >&2\n",
 			wantStderr: "error\n",
+		},
+		{
+			name: "error",
+			cmd: &Cmd{
+				Trace: true,
+				Cmd: &exec.Cmd{
+					Path: "/invalid/path",
+				},
+			},
+			wantErr: true,
 		},
 	}
 
@@ -100,43 +100,4 @@ func TestCmdRun(t *testing.T) {
 			assert.Equal(t, tt.wantStderr, stderrBuf.String())
 		})
 	}
-}
-
-func TestCmdSetTrace(t *testing.T) {
-	tests := []struct {
-		name     string
-		cmd      *Cmd
-		trace    bool
-		expected *bool
-	}{
-		{
-			name:     "set trace to true",
-			cmd:      &Cmd{},
-			trace:    true,
-			expected: boolPtr(true),
-		},
-		{
-			name:     "set trace to false",
-			cmd:      &Cmd{},
-			trace:    false,
-			expected: boolPtr(false),
-		},
-		{
-			name:     "overwrite existing trace value",
-			cmd:      &Cmd{Trace: boolPtr(true)},
-			trace:    false,
-			expected: boolPtr(false),
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.cmd.SetTrace(tt.trace)
-			assert.Equal(t, tt.expected, tt.cmd.Trace)
-		})
-	}
-}
-
-func boolPtr(b bool) *bool {
-	return &b
 }
