+++ type/stringmap.go
