+++ type/stringslice.go
