+++ type/stringslice_test.go
